Divide dialog form into dialog and form
